requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
html5lib==1.1
idna==3.3
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.0.0
